+++ helpers.h
@@ -3,6 +3,6 @@
 
 #include "common_types.h"
 
-void write_uint32(uint32_t number, char *buffer, uint8_t *startIndex);
+void write_uint32(uint32_t number, char *buffer, uint32_t *startIndex);
 
 #endif
